add indexing to baseCommand Class
